requirements.txt
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.41.0
gunicorn==20.0.4
idna==2.8
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==3.2
psycopg2==2.7.2
django-crispy-forms==1.8.1
djangorestframework==3.11.0
pycrypto==2.6.1
Pillow==7.0.0